requirements.txt
opencv-python
pytesseract
openpyxl
rich
requests
Pillow

